Update README.md
I needed to add comments on the last two. Then the description on the first one felt off so I changed it too.

fwiw the working directory in the container will be / by default. I left the leading / off my example because entrypoint.sh will add it when it prepends the (empty) $GITHUB_WORKSPACE.